% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{add_pings}
\alias{add_pings}
\title{Add mentions}
\usage{
add_pings(payload, pings, curr_team)
}
\arguments{
\item{payload}{JSON-like structure for the message body, see https://learn.microsoft.com/en-us/graph/api/chatmessage-post}

\item{pings}{List of emails of Teams users to be alerted}
}
\description{
Add people to be mentioned
}
