% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{make_card_payload}
\alias{make_card_payload}
\title{make_card_payload}
\usage{
make_card_payload(card_items)
}
\arguments{
\item{card_items}{List of objects to go into a card body (see https://adaptivecards.microsoft.com/designer.html)}
}
\description{
Creates message and card wrappers around the content of a card. Designed to
be part of a card creator.
}
