% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{make_project_card}
\alias{make_project_card}
\title{Make project card}
\usage{
make_project_card(run_name, project_name)
}
\arguments{
\item{run_name}{Run name}

\item{project_name}{Project name}
}
\description{
Generate a fancy formatted Teams message describing the outcome of a targets
run.
}
