% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{run_targets}
\alias{run_targets}
\title{Wrapper for running targets}
\usage{
run_targets(
  run_name,
  project_name,
  container_url,
  upload_targets = c(),
  invalidate = FALSE,
  forced = FALSE
)
}
\arguments{
\item{run_name}{Run name}

\item{project_name}{Project name}

\item{container_url}{Azure container URL}

\item{upload_targets}{Name-strings of targets that should be uploaded}

\item{invalidate}{Re-run every target}

\item{forced}{Overwrite blob version}
}
\description{
A complete wrapper for automated/one-touch target runs. Runs tar_make(),
checks for errors, then upload files and sends a Teams message.
}
