% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{send_teams}
\alias{send_teams}
\title{Send Teams payload}
\usage{
send_teams(
  payload,
  pings = NULL,
  channel_name = "Bots Health Check",
  team_name = "Insights"
)
}
\arguments{
\item{payload}{JSON-like structure for the message body, see https://learn.microsoft.com/en-us/graph/api/chatmessage-post}

\item{pings}{Ping users in this message using their emails (case sensitive) as identifiers}

\item{channel_name}{Name of the channel to post on}

\item{team_name}{Name of the team the channel belongs to}
}
\description{
Sends a Teams payload to the specified channel using the Graph API. See
https://learn.microsoft.com/en-us/graph/api/chatmessage-post for details on
the API.
}
\details{
Note that this relies on a user-level authentication. i.e. This can only
post a message where you can post a message.

The payload needs to be in JSON, but we are generating JSON-like structures
in R (using list() to describe dictionaries and arrays) and then converting
it into a JSON string. The content also needs to follow very specific rules.
See make_card_payload() to understand how to attach an AdaptiveCard to a
message, and use the WYSIWYG designer to understand how to create an
AdaptiveCard: https://adaptivecards.microsoft.com/designer.html

If you run into trouble, test with a JSON string that is known to work and
step up from there, e.g.:
'{"body": {"content": "Hello World"}}'

In R JSON-like structure, this is:
list(body = list(content = "Hello World"))
}
