% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{send_teams_message}
\alias{send_teams_message}
\title{Send Teams message}
\usage{
send_teams_message(message_text, ...)
}
\arguments{
\item{message}{Message to send}

\item{channel_name}{Name of the channel to post on}

\item{team_name}{Name of the team the channel belongs to}
}
\description{
Sends a simple Teams message to the specified channel using the Graph API.
See https://learn.microsoft.com/en-us/graph/api/chatmessage-post for details
on the API.
}
\details{
Note that this relies on a user-level authentication. i.e. This can only
post a message where you can post a message.
}
