% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{make_column_items}
\alias{make_column_items}
\title{make_column_items}
\usage{
make_column_items(targ_df, col_name)
}
\arguments{
\item{targ_df}{Dataframe to generate the column from}

\item{col_name}{Name of the channel to post on}

\item{team_name}{Name of the team the channel belongs to}
}
\description{
Make the cells within a column using a dataframe column. Designed to be part
of a card creator.
}
